Rearrange attributes to avoid unnecessary spaces
Unquoted attributes (e.g. ‘foo=bar’) require a space if they are to be
followed by a slash indicating a self-closing tag.  Similarly, due to
a bug in WebKit, unquoted attributes ending with a slash need a space
after them if they are the last attribute.

Rearrange attributes to mitigate those common pitfalls.