Roll ‘has_pre’ checking into build_tag method
Avoid iterating over attributes list multiple times by pushing the
check whether a PRE attribute is present on the tag into the build_tag
method which already iterates over the list.

Furthermore, if keep_pre is False, with this change generating of
a copy of the attributes list is avoided.

This results in some small performance gains:

Before:
    35946    0.227    0.000    0.732    0.000 parser.py:223(handle_starttag)
    35946    0.229    0.000    0.733    0.000 parser.py:223(handle_starttag)
    35946    0.228    0.000    0.730    0.000 parser.py:223(handle_starttag)

After:
    35946    0.208    0.000    0.723    0.000 parser.py:228(handle_starttag)
    35946    0.207    0.000    0.714    0.000 parser.py:228(handle_starttag)
    35946    0.211    0.000    0.723    0.000 parser.py:228(handle_starttag)

Before:
    Ran 82 tests in 0.394s
    Ran 82 tests in 0.397s
    Ran 82 tests in 0.391s

After:
    Ran 82 tests in 0.384s
    Ran 82 tests in 0.380s
    Ran 82 tests in 0.388s